+++ LSL/ViewModels/BarRegionVM.cs
@@ -4,11 +4,7 @@ using LSL.Views;
 using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Reactive.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace LSL.ViewModels
 {
