+++ LSL/ViewModels/InteractionUnits.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Reactive;
-using LSL.Services;
 using LSL.Views;
 using ReactiveUI;
