+++ LSL/ViewModels/RegionalVMBase.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LSL.ViewModels
+﻿namespace LSL.ViewModels
 {
     public abstract class RegionalVMBase : ViewModelBase
     {
