+++ LSL/ViewModels/ServerViewModel.cs
@@ -1,13 +1,11 @@
 ﻿using System;
 using System.Collections.ObjectModel;
-using System.Linq;
 using System.Reactive.Linq;
 using System.Windows.Input;
 using Avalonia.Controls;
 using Avalonia.Controls.Models.TreeDataGrid;
 using Avalonia.Media;
 using Avalonia.Threading;
-using LSL.Services;
 using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
