+++ LSL/ViewModels/ServiceConnector.cs
@@ -1,15 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
 using System.Reactive.Linq;
-using System.Reactive.Threading.Tasks;
-using System.Reflection.Metadata.Ecma335;
 using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
-using Avalonia.Media;
 using Avalonia.Threading;
 using LSL.Services;
 using LSL.Services.Validators;
