+++ LSL/ViewModels/ShellViewModel.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Reactive.Linq;
 using System.Windows.Input;
-using LSL.Services;
 using ReactiveUI;
 
 namespace LSL.ViewModels
